fix(stats): widen ray-count types to 64-bit (Windows u32 overflow) — task-297

+++ src/gui/server_poller.cpp
@@ -104,7 +104,7 @@ void ServerPoller::InvalidateStagedTexture() {
   staged_.has_new_texture = false;
 }
 
-void ServerPoller::SetCalibratedThreshold(unsigned long threshold) {
+void ServerPoller::SetCalibratedThreshold(unsigned long long threshold) {
   calibrated_min_rays_ = threshold;
   calibrated_ = true;
   GUI_LOG_INFO("[Poller] calibration set: threshold={}", threshold);
@@ -185,7 +185,7 @@ void ServerPoller::PollOnce() {
 
       // Quality gate: skip texture overwrite for sparse snapshots (too few rays = visible flicker).
       // Cold start (sim_ray_num == 0) is allowed through — no "old good texture" to preserve.
-      unsigned long min_rays = calibrated_ ? calibrated_min_rays_ : gui::kMinRaysFloor;
+      unsigned long long min_rays = calibrated_ ? calibrated_min_rays_ : gui::kMinRaysFloor;
       bool quality_ok = cached_stats.sim_ray_num == 0 || cached_stats.sim_ray_num >= min_rays;
 
       // Timeout fallback: if quality gate has been rejecting for too long (e.g. empty filter),
